ds_filters.js - module for datasets filters
* config.json(.template) - service config (template)
* resolvers.Annotation.ionImage - fetches image urls from es
* small fixes
