+++ sm_config.json.template
